requirements.txt
keyboard==0.13.5
pillow==10.4.0
plyer==2.1.0
pyautogui==0.9.54
pyperclip==1.9.0
